VIC-3837 BLE: Clear Gatt queues on disconnect
After about 15 or 20 connection cycles, we sometimes ended up in a
state where the next connection would just hang when it should have
been writing the CCC descriptors for the notifiable characteristics of
the cube.  It turned out that this was because we thought that we had
a GATT write in flight and we needed to wait for a callback.  How did
that happen?  Well on the previous connection or connection attempt we
had queued up a GATT write for a descriptor or characteristic and then
got disconnected.  But, we never erased the items for that connection
from the queue.  In addition, we didn't reset the flags that tell us
we are clear to transmit a GATT write or GATT read.  With this change,
we do a proper cleanup on disconnection.

(cherry picked from commit 40879c59a543811381db8ae4f6156300af22aa1c)
Signed-off-by: Stuart Eichert <seichert@anki.com>