VIC-4180 BLE : Deregister for notifications
When we connect to a cube, we register for notifications from the
characteristics that support them.  However, on disconnecting from a
cube, I assumed (always dangerous) that Fluoride would clean up our
registrations.  I was wrong, it doesn't.  In fact, there is a fixed
size table inside Fluoride for the registrations.  The maximum number,
BTA_GATTC_NOTIF_REG_MAX, is 15.  This is defined in
bta/gatt/bta_gattc_int.h inside Fluoride if you are
interested. Eventually, we run out of space in the table and when
trying to connect to a cube, we will fail to register for
notifications with an error indicating that we are out of resources.
For our cubes, I believe they each have 2 characteristics that support
notification.  Therefore we would be able to connect to 7 cubes, but
on the 8th one we would fail.  However, we would be able to reconnect
to one of the original 7 cubes, because our registration for
notifications was already in the table.

To fix this, on disconnect, whether we initiate it or Fluoride tells
us about it, we will deregister for notifications from every
characteristic that we registered for.