BI-995 Remove 30_set_submodule_branch.sh
This script was used to check out the `master` branch of the
`anki/victor` submodule on TeamCity during a build of Vicos -> Dev
Unstable.  I changed the build on TeamCity to directly fetch the
`master` branch of `victor.git` into the `anki/victor` sub directory.
This script is no longer needed.