VIC-3132 Create robot private key + gateway ca cert
vic-gateway communication uses TLS, which requires a self-signed CA cert
that will be sent to clients and cerified by vic-gateway during the connection
handshake. The robot private key is now created during the mount-data step
and stored in `/data/etc`.  The gateway cert is generated (if necessary) before
vic-gateway starts.