VIC-4712 Restore vic-christen for setting robot name
We want to set the robot name + hostanem early during boot, so that it is available before
WiFi starts and before other `vic-*` processes run.  This will ensure that the robot name and
hostname name are set before anything might need them.

We currently need the robot name / hostname to be set before vic-switchboard so that the
vic-gateway process can access it and so that the self-signed CA cert, which requires the
robot name as the common name (domain name) can be created.

This change reverts:

Revert "Remove vic-christen now that the functionality lives in vic-switchboard."

This reverts commit 968fba17b2794d51663b61c06ebf3c9a178cc509.

Conflicts:
	anki/victor
	poky/meta-qti-bsp/recipes-core/ankiinit/files/ankiinit.sh

In addtion, the original `vic-christen` code now uses the switchboard block partition to
read/write data. It is uses a minimal data representation for accessing name data that
is compatible with all versions of `vic-switchboard`.