VIC-4547 socket-activated fault-code-handler
This changeset provides a separate service for receiving fault codes generated
by programs running on the robot and taking performing useful actions.  This removes
the need to do error handling inside `vic-anim` and also makes it easy to take
additional steps (such as auto-restart or collecting log dumps).

fault-code-handler is configured via systemd to be run in response to data written
to a fifo socket at `/run/fault_code`. Once activated, it stops all services in the
anki-robot.target (victor.target), displays the fault code and optionally restarts
the target after a configurable delay. The optional systemd `.env` file lives in
`/anki/etc` so that we can couple restart and timing behavior with victor programs.

Note that this is a shell script instead of a C program because its main function is
to call other programs. In the future, we should consider replacing the bash `read`
built-in with a C program that reads input and validates the fault code by checking
it against the values defined in `faultCode.h` (in victor repo).