VIC-4430: Integrate NetworkManager into vicos
This patch replaces connman with networkmanager as part of the vicos build. The
additional libraries libndp, libgudev, and gnome-common are required by
networkmanager.

In order for NetworkManager to save it's connection profiles correctly this
patch also installs a soft link from /etc/NetworkManager/system-connections to
/data/etc/NetworkManager/system-connections.

NetworkManager is also configured to manage /etc/resolv.conf which is now
linked to /var/run/NetworkManager/resolv.conf. This configuration file will be
automatically be updated by NetworkManager when connected to a new device.

The pre-down.d/0-udhcpd pre-up.d/0-udhcpd scripts are called by nm-disptacher
when connection status changes. These scripts are used to enable/disable the
dhcp server as well as set the power level of the wlan device.