VIC-5186 Wait for automatic OTA to finish random sleep
For the hourly automatic OTA, we have systemd randomly sleep up to
1800 seconds before launching /anki/bin/update-engine.  We do this to
avoid having all the robots hit the OTA server at the same time.
Unfortunately, this meant that often we went past systemd's internal
timeout for starting a service.  To fix this, I updated the service
definition to ask systemd to give us 1830 seconds to start.