VIC-4820 Fix anki-audio-init dependencies (#326)
The root cause of this issue is that `anki-audio-init` sometimes starts before the alsa device entries are present in `/dev/snd`.  This results in `amixer` commands failing.  The correct way to solve this seems to be with udev + systemd, but I am struggling to get systemd talking to udev for the purpose of correctly setting the status of a `.device` file to be used as a dependency.  This fix is required for use in beta builds, so adding back an explicit loop/wait in anki-audio-init should fix the problem until we have time implement the proper udev/systemd solution.

Other supporting improvements:

- Add alsa-utils so that we can stop using tinyalsa / tinymix tools

- User amixer command in anki-audio-init

- Make init_audio a oneshot service
  Needs to be oneshot + RemainAfterExit in order to complete before other services start

- Add a dependency on sound.target
  Based on the current `udev` rules, `sound.target` should activate when the sound device is available. We should be able to use this to "wait for sound to be ready", but it doesn't seem to work.