VIC-5325 Build "userdev" version of victor
"userdev" is the shipping version (aka "user") with a few minor
things for developers to be able to troubleshoot and diagnose
bugs.  "userdev" builds should autoupdate to "userdev" builds.  To
help make that happen, I added a new build script to victor for
building the "userdev" variant.  This change to victor.bb calls
that script.

Note:  As of this point, the "userdev" version of victor is the
"release" version instead of the "shipping" version.  That will change
in a future commit to the victor repo.