VIC-5811 Customize connman for victor
These are changes to enable running connman-1.36 in our current OS.
The default configuration runs, but either fails due to dbus access permissions
or failure to load detect kernel support for iptables. This is present, but it appears
that some aspect of modprobe is broken, causing the checks to fail and crash the daemon.

Also, because we need to manage WiFi without human intervention, there are a few WiFi settings
in  connman related to auto/background scanning that we might need to change from the hard-coded
defaults.

- Install dbus rules for connman access

- Restore change to set connman conf dir to /data

- Patch connman to skip calling xtables_insmod

In a connman 1.30, this call always failed and execution continued. In 1.36, the call
causes connman to crash with a SEGV.  It does not appear to affect our current functionality,
so we leave it off.

- Add default main.conf file for conmann

- Add tunable WiFi settings for connman scanning