VIC-1951 Create system user/group for logd
We want to allow processes, like ankibluetoothd, to run as a regular
user but still be able to write to logcat.  This change will make it
so that /dev/socket/logd* are owned by system:system and writable
by members of the "system" group.  Apps, not running as root, that are
going to write to logcat need to be a member of the "system" group.

(cherry picked from commit 96ca6b4dbbc95435fdcf84fb2c45f45ae8e7c518)
Signed-off-by: Stuart Eichert <seichert@anki.com>