VIC-4615 Camera stop stream hanging (#347)
Ensure raw frame callback doesn't try to do anything once camera_stop() has been called by adding a shutdown_lock mutex that remains locked while the camera is off.

Update buffer book keeping when the VIDIOC_STREAMOFF ioctl call is made. This removes and unlocks all buffers from the kernel so we need to update our book keeping to match.

When stopping the stream we wait for two command threads to join. Sometimes the threads do not stop and we would block forever on the join. Switched to using a timed join with a timeout of 60 seconds. If the join times out then cancel the threads.