VIC-3034 Encrypt /data partition
Encryption makes it harder for an attacker to get the user's data and
snoop around the saved state of our various apps.  Attackers will
certainly dump the contents of the eMMC chip shortly after the robot's
release to the public.  We don't want to make it any easier for them
than necessary.

This change does NOT use Qualcomm's TrustZone for storage/retrieval of
the passphrase.  Instead, the passphrase is randomly generated,
obfuscated and then stored on the robot's eMMC.  If a motivated
attacker can figure out the obfuscation technique, they will
eventually be able to decrypt the /data partition.  Later, we hope to
add a change to use TrustZone for the passphrase, so that only the
robot will be able to decrypt the /data partition.

In this change, we use the `cryptsetup` program with `LUKS` to format
and open the encrypted volume for `/data`.  We chose LUKS over plain
dm-crypt to avoid making mistakes that would weaken the security of
this encrypted volume.  More information on `cryptsetup` and `LUKS`
can be found at `man cryptsetup` or on the cryptsetup FAQ at:
https://gitlab.com/cryptsetup/cryptsetup/wikis/FrequentlyAskedQuestions

Credit where credit is due.  99% of this work was done by Alexander
Kaplan <alexander@anki.com>.  My work was just to clean up a few loose
ends and prepare this work for committing into the code base.
Mistakes should be blamed on me, not him.