VIC-3557 Tag dev unstable build based on OS version
Previously, we were tagging the dev unstable build based on the
version from the anki/victor repository.  However, in our build
scripts, we were creating the OTA_VERSION based on the version from
the OS.  In other build scripts, we assumed the OTA_VERSION and the
tag name were the same.  That was bad because it is possible for one
of them to have their version bumped without the other one changing.

This change makes the tag name match the OTA_VERSION.  We still need to
think more about how we create our version numbers, but this should be
a safe change that will stop the Promote To LKG script from breaking.