VIC-6069 Write /run/data_cleared after clearing data
When we boot up, we want to know if the user just did a "Clear User
Data".  With this change, when the user does "Clear User Data", we
will write the word "WIPE" to the start of the data partition.  On the
next boot, if we see the word "WIPE" at the start of the data
partition, we will write out the `/run/data_cleared` file so that our
apps will know they are booting up after the user cleared their data.

I also tested doing a second reboot to make sure that
`/run/data_cleared` did not show up after formatting and mounting the
data partition after the clear.