VIC-5603 Add persistent firewall service
This is a simple systemd service to configure iptables/ip6tables to block
all incoming packets accept for services that we expose. In shipping (user) builds,
this is limited to port 443. In userdev and eng builds, we allow additional ports for
debugging. In the future, it might make more sense to allow part of this config to
live in the `victor` repo or use some kind of template pattern for the rules file
to allow base + userdev/eng additions.  The intention here is to create a simple
solution for 1.0.