VIC-6486: Update victor.bb to clean source directory
The source directory is used as the build directory when using the externalsrc module. This patch updates the recipe to ensure build artifacts are fully cleaned from the source directory. This patch also presumes that build artifacts are added to the .gitignore configuration. This will ensure new files are not deleted by the clean step unless they are also added to the .gitignore configuration.