VIC-5601 vic-christen: Validate name field
Previously, when we read the name field from the switchboard
partition, we assumed that it matched our required format for the
robot's name.  Initialization scripts would then use this name
unchecked and potentially expose the robot to a shell injection attack
or other threat.  In the event that an attacker has modified the
switchboard partition and corrupted the name, we want to detect it and
generate a new compliant name.  This change also guards against an
attacker appending non-null bytes to the end of a valid name.