VIC-1951 Backport AmbientCapabilities support from systemd v229
Ambient capabilities are required to set the capabilties of a process exec'd as a
non-privleged user, where the caps are not set on the binary. In short, setting
`AmbientCapabilities` in a systemd service file allows us to easily start our processes
as non-privileges users with the minimum required capabilities. The key feature is that we
do not need to specify caps on the files, which would be difficult for our current deployment
solution. ambient capabilities in general were added to the kernel in 4.7, but were
backported to our 3.18 version.

(cherry picked from commit 08254493bf27b67030cecb2a2b7a05b3b678dfe5)
Signed-off-by: Stuart Eichert <seichert@anki.com>