VIC-1951 Support non-root users for vic-* services
* pre-create `/data` dirs required by vic- process

`/data` dirs that are used by vic-* processes are now provisioned after creating
the userdata partition, so that we can ensure that the permissions and ownership
credentials are set correctly.

* Specify user/group/caps in vic-* service files

`vic-gateway` now runs as `net:anki`.  It is allowed to create & bind a socket to port
`443`.  The `anki` group ownership is used mainly for access to files and domain sockets
for IPC.  For now other processes are assigned to the `anki` group, with a `umask` such
that files (and domain sockets) they create will be available to other members of the `anki`
group.  This should allow us to slowly convert process to run without root privileges without
breaking IPC. `AmbientCapabilities` defined in `systemd` service units allow us to change
capabilities of the process owned by a non-privileged user.

* user/group/perm cleanup for /anki/vic-*

The inet group is used by the ANDROID_PARANOID_NETWORKING patch to check whether
a non-system (non-root) user is allowed to open a network socket.

Related:

- Change the cloud uid:gid to 888:888 in order to match the ownership of existing
cloud files in the `factory` partition.

- Ensure vic-* executables (mainly vic-gateway) has the correct ownership so that
it can be executed by the net user.