VIC-6439 Temporarily allow vic-gateway to read any file
This fixes an issue where `vic-gateway` is silently failing to handle tokens
because it is attempting to access a file that is does not have permission
to read when running as `net`.

I was able to repro the issue with build 1677 and narrowed down the required
capabilities to just adding CAP_DAC_READ_SEARCH. I suspect the issue might be
related to attempting to read the ESN directly from the block partition, but
I have not isolated the problem to that specific case yet.