VIC-6387 Use TrustZone for encrypting '/data'
Previously, we generated a random passphrase to encrypt the '/data'
partition.  We then xor'd that passphrase with the serial number of
the eMMC and stored it in the switchboard partition.  This is security
through obscurity.  If an attacker figures out how to transform the
passphrase into its original form, he can use that to decrypt the
'/data' partition outside of the robot.

This change uses Qualcomm TrustZone's KeyMaster trustlet to generate
an RSA key blob.  We then store that RSA key blob on the switchboard
partition and ask TrustZone to use it to create a digital signature
of a constant array of bytes.  This digital signature is 256
hexadecimal characters.  We use that as our new passphrase for the
LUKS encrypted '/data' partition.  Without this robot's specific
TrustZone, it is not possible to create the passphrase used to decrypt
the '/data' partition.