VIC-6770 Make sure /dev/socket/logd* are owned by system
There is a race condition with how logd starts up.  Sometimes,
a thread or process running as root will create the /dev/socket/logd*
unix domain sockets.  Other times, after a different thread/process has
dropped privileges and done a setuid/setgid to the system user the
sockets will be created.

We want the sockets to be owned by the system:system user/group
so that only root and members of the system group can write to logcat.

This change doesn't fix the root cause, but instead adds a step
to the initialization script to make sure the sockets have the
correct ownership.