VIC-1951 Run vic-cloud as cloud user
To run as the cloud user, we also currently add the `CAP_DAC_READ_SEARCH`
capability to work around the fact that `/factory/cloud` has the wrong
mode (440), preventing us from "searching" its contents. The alternative
would be detecting it, remounting rw, fixing modes, the remounting ro, which
seems risky (we never mount this partition rw outside of the factory).