VIC-7043 ota/Makefile - Support reboot_after_install
By default, we don't automatically reboot after an install of an OTA.
However, in the future, we may have to release an emergency OTA that
should go into effect immediately.  The reboot_after_install option
allows this.  There is a complementary change in update-engine to
reboot the robot after successfully installing OTAs that use this
option.