VIC-7146 Reboot Stresser Service
We are concerned that there may be a bug with mounting the `/data`
partition at startup that results in it getting reformatted and the
user losing their data.  This stress test is meant to be run hundreds,
if not thousands of times in a row, to prove that the robot can boot
up and mount its `/data` partition without error.  If the `/data`
partition is reformatted, the robot will lose its WiFi credentials and
we will no longer be able to ping it or ssh into it.  If that happens,
we know that the test failed and we need to investigate further.

This service and its timer will only be in development builds, not in
user (aka shipping) builds.  By default, this service and its
associated timer are disabled.  To enable it and start stress testing:

```
$ mount -o rw,remount /
$ systemctl enable reboot-stresser.timer
$ reboot
```

To disable it:
```
$ mount -o rw,remount /
$ systemctl disable reboot-stresser.timer
$ reboot
```