VIC-7326 Do not start qseecomd or sfs_config
It is possible, but unproven, that `qseecomd` is interefering with
`user-data-locker` 's use of TrustZone.  From reading the logs,
it seems that it runs at more or less the same time.  There may be a race
condition where that causes `user-data-locker` to fail to get the
passphrase to decrypt the `/data` partition.

sfs_config is unused on our board, so removing it.