VIC-7326 Make mount-data.service require /firmware
When `mount-data.service` runs, `user-data-locker`, asks Qualcomm's
TrustZone to load the KeyMaster firmware image.  It gets this image
from the `/firmware` mount point.  While I don't have any evidence to
suggest that `/firmware` was ever NOT mounted during the start of
`mount-data.service`, this adds an explicit requirement for it just to
be safe.