VIC-7326 Add more logging if we fail to mount `/data`
I don't know why we sometimes fail to mount `/data`.  With this change,
we will more carefully check the return code from each operation we
perform while trying to mount `/data`.  If we fail and have to reformat
it, we will write out `/data/format-data.log` with kernel messages and
debug messages from `journalctl`.  Hopefully that will provide a clue
as to why we failed.