VIC-7508 make_delta_ota.py : Support reboot_after_install
As noted in VIC-7043, we may want to deploy an OTA file that should
result in the robot rebooting immediately after install.  This may be
important for security critical releases.  With this change, when we
make delta OTA files they will have a reboot_after_install field set
as follows:

1. Take the value from the --reboot-after-install command line
2. If no command line value, use the value from the "new" OTA file's
   manifest
3. If no value in manifest, use 0