VIC-8086: Disable msm dload to fix halt on panic
This commit disables the msm dload option in the victor debug kernel config. The
msm dload system causes the bootloader to halt on restart when magic values are
found in imem written by the panic handler. This is used by qcom to debug panics
where they can attach and download the contents of memory very early in the boot
sequence.