VIC-8849 Allow vic-gateway to reboot the robot
vic-gateway handles a message that allows a client to check for an update and reboot if
one is ready to be applied.  In order to reboot, vic-gateway required CAP_SYS_BOOT or another
mechanism of signalling a privileged process to reboot. This change adds CAP_SYS_BOOT and since
`kexec` is turned off (also enabled by CAP_SYS_BOOT), this capability only allows reboot.