VIC-10785: Always create userdata in mount-data
The make_cryptdata function after preforming the cryptsetup mounts the
/data partition. The "mounted" result was not being set resulting in the
make_userdata function never being called causing issues when user data
was wiped. Really we always want to try to create the userdata as it's
not distructive. This patch also renames the make_cryptdata to
make_and_mount_cryptdata to make it more clear what the function does.