VIC-6790 Drop BLE connection to cube if vic-engine cancels
We were getting into an unexpected state where the client,
`vic-engine`, was cancelling an in-progress BLE connection to a cube.
We thought we had thoroughly instructed the Fluoride stack to cancel
the connection, but sure enough a couple of seconds later we would get
a callback indicating that the connection was now complete.  This
caused confusion for `vic-engine`, which didn't think that the robot
was connected to a cube.  With this change, we introduce a new state,
Disconnecting, to keep track of the fact that the client doesn't want
to connect to this cube.  If we later get a callback from Fluoride
telling us that we are connected we will try to thoroughly disconnect
again.  From my testing, that seems to work.

(cherry picked from commit 5efff3415ce43bb1e8491bf2abde0ffa7ec41dfe)
Signed-off-by: Stuart Eichert <seichert@anki.com>