VIC-11271 Wait 0 to 60 minutes before downloading OTA
We check for an OTA download once an hour.  Previously, we delayed a
random amount up to 30 minutes before trying to download the OTA.
That meant we were hitting the CDN all in the first half an hour of
every hour.  It is better to spread out our CDN access to the full
hour.