VIC-9820 Integrate lttng kernel and userspace tracing
The Linux Trace Toolkit (lttng) is a lightweight tracing system that
supports both kernel and userspace traces. This commit adds the
ankitrace yocto package to the dev and userdev images. The ankitrace package
provides support for starting stopping and recording a lttng trace. The trace
events have been selected to capture most events while trying to keep the
total event rate down. The target is to capture around three to four seconds of
trace data in a record event.

The fault-code-handler has been updated to trigger an ankitrace record which
will store the trace record in /data/ankitrace/faults/CODE/ankitrace-DATETIME.tgz.