VIC-11362 /sbin/wipe-all-wifi-configs
Chewie or `mac-client` can ask `vic-switchboard` to delete ALL of the
WiFi configurations on the robot.  Currently, with `vic-switchboard`
running as root, it does this by enumerating the configuration
directories below `/data/lib/connman` and then deleting them.  This
won't work for running as non-root due to the permissions of those
directories.  Instead, this new script, `wipe-all-wifi-configs` can be
run with `sudo` to delete all of the configurations and restart the
`wpa_supplicant` and `connman`.

(cherry picked from commit ca994525617c53cd69de47bec2545b5fdb2c3866)
Signed-off-by: Stuart Eichert <seichert@anki.com>