VIC-11681 Split record and archive functions in ankitrace
When capturing a trace in fault-code-handler ankitrace could take over
thirty seconds to compress the trace to /data. This commit splits the
record and archive functions in ankitrace. With this change
fault-code-handler will record a trace as soon as possible in order to
get the data as close to the fault event as possible. The
fault-code-handler then displays the fault code and finally copies the
trace to /data for later retrieval. This commit also removes the
compression step as it was taking almost thirty seconds to complete.