VIC-11799 Move clock forward after power outage
After a power outage, the robot's hardware clock will reset to 0 and
then start ticking seconds after the power comes back on.  We can
safely add these seconds to our saved time to get a more accurate wall
clock time until we get a response from an NTP server.  This will help
for short power outages like when the user chooses to power cycle the
robot by pressing and holding the button.  Unfortunately, it won't
really matter if the power outage lasts for an hour.