VIC-11712 Remove journald to reduce memory usage
Journald is using over 13MB of ram as well as additional memory in the
pagecache which puts a lot of pressure on the system. This patch
disables journald and configures systemd to forward messages to kmsg in
place of journald which are then picked up by syslog. When reading the
systemd messages from kmsg they appear as if they are kernel messages by
default. To format these logs correctly this commit adds a rewrite step
to syslog-ng which changes the program to "systemd" if it is a pid 1
log.