VIC-11834 Enable network-online.target via connman
We want to try and synchronize time from the Internet as soon as the
network comes online.  This change adds the `connman-wait-online`
service.  `systemd` uses this service to determine if the
`network-online.target` is active and therefore it can activate units
that depend on `network-online.target`.  A later commit will add a
service that waits for `network-online.target` and then immediately
tries to sychronize time with the Internet.

Our NTP client, `chronyd`, does NOT need this service.  `chronyd`
monitors the state of the network on its own and will start trying to
synchronize time from NTP servers as soon as it can.