VIC-11799 Do not move clock into future at boot
This is a rare and possibly developer only scenario where the clock
could get set to a time in the future before being corrected by NTP or
tlsdate:

1. Have robot sitting on charger and not powering off for 3 or 4 days
2. Install build with /etc/timestamp that is only 1 day old
3. Boot robot

After the `fake-hwclock` service runs it will first set the time to
match the timestamp in `/etc/timestamp`, but then will add the number
of seconds from the hardware clock (which is 3 or 4 days worth) to the
current time.  As a result the clock jumps into the future.  Later
(usually within 10 to 15 seconds) when NTP synchronizes time, it will
move the clock back to the real time.

The fix for this is to only add in the seconds from the hardware clock
if there is a saved /data/opt/fake-hwclock/fake-hwclock.data.  If we
have a saved /data/opt/fake-hwclock/fake-hwclock.data and the value of
the hardware clock is greater than what is saved, we only add in the
difference.  If the hardware clock value is less, then we add in the
absolute number of seconds because we know that this indicates a power
outage and the hardware clock having reset to 0 since we last booted
the robot.