VIC-11834 httpsdate - Set clock via HTTPS on boot
Our applications can benefit from having an accurate clock as soon as
possible during bootup.  Currently, we rely on our NTP client,
`chronyd`, to synchronize the clock.  In my testing, it takes
`chronyd` about 12 seconds after we gain network access to do this.
With httpsdate, it takes about 2 seconds.  This change does the
following:

1. Do an HTTPS HEAD request to www.anki.com and get the date/time from
   the `date: ` header.
2. Set the clock to this date
3. Write out the file `/run/https.timesync` so that other processes
   will know that the time has been synchronized
4. Emit a DAS event so we can monitor how will this feature works

This change does NOT disable `chronyd` or interfere with it in any
fashion.  There is no guarantee that this HTTPS method will work.  We
still need NTP to regularly and reliably synchronize our clock.