VIC-11364 Allow vic-engine to set CPU speed
When `vic-engine` runs as the `engine` user, it will need to be able
to set the CPU speed.  This change makes it so that `vic-engine` is
the owner of the relevant files in the sysfs for doing this.

(cherry picked from commit 3599500e4b8c7acc5682c6a6934000d0ac0aad60)
Signed-off-by: Stuart Eichert <seichert@anki.com>