VIC-11712: Increase limit of per cpu pagelist size
Sets the fraction of total pages that the per cpu pagelist can have
before it gets flushed to the buddy alocator. The per cpu pagelist is
used to prefer "hot" pages however it is only used for single pages.
This increases the amount of memory available for page allocations in
the hot path but it should also reduce memory fragmentation as larger
chuncks of memory won't have to be carved up for single page
allocations.