VIC-11364 Allow 'vic-engine' to talk to the camera
When `vic-engine` runs as the `engine` user it will need to be able to
talk to the camera service.  These changes make it so that the camera
service is available to users in the `camera` group.

(cherry picked from commit 7ee792a6fd06a030d15a87a0d095e36526443f9e)
Signed-off-by: Stuart Eichert <seichert@anki.com>