VIC-11364 write to /tmp/data_cleared instead of /run
When `vic-engine` runs as the `engine` user, it checks for the
`data_cleared` file and if it is found, it emits a DAS event and then
deletes the file.  The `engine` user does not have permissions to
delete a file under `/run`, but it can delete a file that it owns
under `/tmp`.