VIC-11364 Set perms and owner for /anki/*
When `vic-engine` was running as the `engine` user, it wasn't able to
access the files under `/anki/data` and `/anki/etc` on the robot due
to the directories and files being owned by `root:root` and not being
readable/searchable by the `engine` user.  This change corrects that
problem and goes a little further to tighten up the permissions on all
the files under `/anki` and set the owner and group appropriately.
Most files will be owned by `anki:anki` with some overrides to have
certain executables owned by the user who is expected to execute them.