VIC-11363 Run vic-robot as 'robot'
For security, we want to run `vic-robot` as the `robot` user instead
of `root`.

*Note*
Had to add `ExecStartPre` line to `chgrp anki` for the `emr`
partition like we do in `vic-engine.service`.  This is because
this version of `vic-robot` reads the EMR at startup.

(cherry picked from commit 4d2e05cc292db1276dc4e4f634cd70dc468b504d)
Signed-off-by: Stuart Eichert <seichert@anki.com>