VIC-12426: Reverting removal of logwrapper
Logwrapper forwards stdin to logd. This is an expensive process however
after removal the stdout from processes are being forwarded to journald
which sometimes decides to buffer the data until the service is stopped.

These changes will be reverted until the root cause of the issue can be
fixed.

Revert "VIC-12283 update-engine.service: Do not use logwrapper"
Revert "VIC-11712: Remove the use of logwrapper in systemd services"
 revert commit cd3a0dd62d2478896065a877aa2c607c3581a5b4.
 revert commit b9739b6392d63959a5370760171131c3e451a6e7.