VIC-12519 Reboot overnight if an OS update is pending
We have a few things that typically block us from doing an overnight
reboot:
* not being in powersave mode
* insufficient uptime (less than 4 hours)
* being slightly after 5am

If an OS update is pending, we really don't care about any of these.
The only thing that should stop us is the presence of the inhibitor
files: `/data/inhibit_reboot` or `/run/inhibit_reboot` . Their
presence would indicate that one of our processes or a QA test really
doesn't want the robot to do the overnight reboot.