VIC-12601: Remove memlock limit for vic-robot
The default ulimit for locked memory is 65536 bytes. The addition of
lttng and taskExecutor pushed vic-robot over the limit as they both add
additional threads to vic-robot. When a process uses mlockall with
MCL_FUTURE and reaches the limit any additional allocations that trigger
pages to be mapped in will result in a failed allocation. This result is
a system that appears to be out of memory. This commit addresses the
issue by removing the memlock limit on vic-robot.