VIC-12963 update-os: Handle empty string crash
Sometimes, when trying to present progress to the user, we would hit a
case where the `/run/update-engine/progress` file was empty.  This
resulted in a crash when trying to calculate the percentage complete.
It looks like this:

```
Updating to  1.4.0.2665d ( 45% )/sbin/update-os: line 81: 100 *  / 4791
: syntax error: operand expected (error token is "/ 4791 ")
```

With this change, we check to make sure that we have non-zero length
strings before trying to calculate a percentage with them.