VIC-12963 update-os: Support delta-latest
Sometimes you just want to try a delta update to the latest and
greatest.  Some of our QA team has really really slow Internet and it
takes forever for them to download a full OTA to the latest version.  A
full OTA can be around 200MB.  By contrast a delta OTA could be as small
as 10MB and probably not more than 40MB.  With this new feature, anyone
can do:

```
$ update-os delta-latest
```

to do a delta OTA update to the latest version.

It won't work in the following cases:
1. You have deployed over your installed OS with victor artifacts.
   This corrupts the rootfs hash and makes it impossible to do a delta.
2. You are already on the latest
3. Your OS build is more than 10 versions behind the latest. In this
   case, you will likely upgrade to a newer version and then have to try
   again.  Possibly still faster than doing a full OTA if you have
   really slow Internet.

It is expected that this feature will be used on robots that are running
a build from the `master-dev` branch.  However, you can try it on
`rc-dev` if you wish.