VIC-4461: Move msm_serial_hs rx/tx processing to isr context
This patch removes the workqueue used to offload msm_serial_hs rx/tx processing
from the sps isr. After profiling this section it was shown that queuing work
was in some cases taking much longer than it was taking to finish the work being
queued. This change reduces the maximum latency in the rx path by 170us in the
current configuration.

Signed-off-by: Mathew Prokos <mprokos@anki.com>