VIC-4461 msm-3.18: Turn on fully dynamic ticks
The current configuration is set to dynamic tick on idle. This means that the
kernel can eliminate tick interrupts when the core is in idle state. This is
nice for saving power however our system does not idle very much. Changing this
to full dynamic means that all cores except core 0 remove the periodic tick
interrupts. This works because while we have vic-robot set as SCHED_FIFO and
other tasks set as SCHED_OTHER the dynamic tick kernel can determine that sched
ticks are not needed on that core which eliminate jitter in our RT application.