VIC-4461: Remove core 3 isolation
This patch removes the core 3 isolation and removes cpu affinity settings on the
vic-robot process. This has slightly decreases the number of "too long" ticks in
vic-robot and slightly increases the maximum latency between vic-robot ticks.

This change is possible because other previous changes in this series have
reduced the scheduling latency enough that isolating core 3 no longer provides
much benefit. This given with the fact that core 3 averages around 30% load
means we can pull a little more performance out of the system if we allow other
processes to use this core.