VIC-4461: Always enable low_latency on TTY devices
The TTY low_latency patches required the ASYNC_LOW_LATENCY flag to be set on the
TTY device. Setting this flag however requires CAP_SYS_ADMIN which negates many
of the security benefits of running our processes without root privileges. This
patch changes the behavior to just always use the high priority kthread so the
flag is not needed.