VIC-11830: Upload debug log in response to fault code (#589)
Revise fault-code-handler to capture recent log messages when a fault occurs.
Log messages are created in "outgoing" directory to be uploaded by vic-log-uploader.

Fault code handler checks for ON_FAULT_UPLOAD_LOG=1. If this flag is not set, logs are not captured.

Fault code handler checks for special file /run/das_allow_upload.  If this file is not present, logs are not captured.