VIC-13292 Let vic-engine read/write gpio0
On Whiskey, `vic-engine` will need to read and write `gpio0` for the
time-of-flight sensor.

This change re-factors and simplifies the `udev` rules for `gpio` to
only change the ownership of the relevant files to the `engine` user.
The rest are left as owned by `root`.

In addition, for security reasons, the `engine` user will no longer be
allowed to export GPIO pins.  Instead `gpio94` and `gpio0` will be
exported by `systemd` as the `root` user when starting the
`vic-engine.service`.