VIC-12144 Generate canned fs config in victor.bb
In ed1091c84b, I added additional scripting to generate the canned
filesystem config for files under `/anki/` on the robot.  I put this
scripting in `apq8009-base-image.inc`.  It is more appropriate to define
it in `victor.bb` where we already define the list of users and groups
that we wish to add to the robot.