VIC-13084: Change ownership of systemd-notify
Because of the way systemd-notify works it can't be run under a sudo command to
send signals to systemd. The --pid argument seems like it should work however I
have not had luck. This patch changes the ownership in the victor install script
so we don't have to use sudo.